[IMP] sale_blanket_order: added Route to sale_blanket_order
