+++ sale_blanket_order/__manifest__.py
@@ -5,7 +5,7 @@
     "category": "Sale",
     "license": "AGPL-3",
     "author": "Acsone SA/NV, Odoo Community Association (OCA)",
-    "version": "15.0.1.2.0",
+    "version": "16.0.1.0.1",
     "website": "https://github.com/OCA/sale-workflow",
     "summary": "Blanket Orders",
     "depends": ["uom", "sale_management", "web_action_conditionable"],
@@ -17,6 +17,7 @@
         "wizard/create_sale_orders.xml",
         "views/sale_config_settings.xml",
         "views/sale_blanket_order_views.xml",
+        "views/sale_blanket_order_line_views.xml",
         "views/sale_order_views.xml",
         "report/templates.xml",
         "report/report.xml",
