+++ sale_blanket_order/models/__init__.py
@@ -1,3 +1,5 @@
-from . import blanket_orders
-from . import sale_orders
+from . import blanket_order
+from . import blanket_order_line
+from . import sale_order
+from . import sale_order_line
 from . import sale_config_settings
