+++ sale_blanket_order/views/sale_blanket_order_line_views.xml
@@ -1,0 +1,181 @@
+<?xml version="1.0" encoding="utf-8" ?>
+<odoo>
+
+    <!-- VIEWS -->
+
+    <record id="sale_blanket_order_line_tree" model="ir.ui.view">
+        <field name="name">sale.blanket.order.line.tree</field>
+        <field name="model">sale.blanket.order.line</field>
+        <field name="arch" type="xml">
+            <tree create="false">
+                <field name="sequence" widget="handle" />
+                <field name="name" invisible="1" />
+                <field name="order_id" />
+                <field
+                    name="product_id"
+                    context="{'partner_id':parent.partner_id, 'quantity':original_uom_qty, 'company_id': company_id}"
+                />
+                <field name="product_uom" invisible="1" />
+                <field name="price_unit" />
+                <field name="date_schedule" />
+                <field
+                    name="original_uom_qty"
+                    string="Original Qty"
+                    context="{'partner_id':parent.partner_id, 'quantity':original_uom_qty, 'company_id': company_id}"
+                />
+                <field name="ordered_uom_qty" />
+                <field name="invoiced_uom_qty" />
+                <field name="delivered_uom_qty" />
+                <field name="remaining_uom_qty" />
+                <field name="company_id" invisible="1" />
+            </tree>
+        </field>
+    </record>
+
+    <record id="sale_blanket_order_line_form" model="ir.ui.view">
+        <field name="name">sale.blanket.order.line.form</field>
+        <field name="model">sale.blanket.order.line</field>
+        <field name="priority" eval="20" />
+        <field name="arch" type="xml">
+            <form string="Sale Blanket Order Line" duplicate="false">
+                <sheet>
+                    <h1>
+                        <field name="order_id" />
+                    </h1>
+                    <group>
+                        <group>
+                            <group>
+                                <field name="product_id" readonly="1" />
+                                <field name="price_unit" readonly="1" />
+                                <field name="partner_id" readonly="1" />
+                                <field name="date_schedule" readonly="1" />
+                                <field
+                                    name="company_id"
+                                    groups="base.group_multi_company"
+                                    optional="show"
+                                    options="{'no_create': True}"
+                                />
+                            </group>
+                        </group>
+                        <group>
+                            <label for="original_uom_qty" />
+                            <div>
+                                <field
+                                    name="original_uom_qty"
+                                    class="oe_inline"
+                                    readonly="1"
+                                />
+                                <field
+                                    name="product_uom"
+                                    groups="uom.group_uom"
+                                    class="oe_inline"
+                                    readonly="1"
+                                />
+                            </div>
+                            <label for="ordered_uom_qty" />
+                            <div>
+                                <field
+                                    name="ordered_uom_qty"
+                                    class="oe_inline"
+                                    readonly="1"
+                                />
+                                <field
+                                    name="product_uom"
+                                    groups="uom.group_uom"
+                                    class="oe_inline"
+                                    readonly="1"
+                                />
+                            </div>
+                            <label for="invoiced_uom_qty" />
+                            <div>
+                                <field
+                                    name="invoiced_uom_qty"
+                                    class="oe_inline"
+                                    readonly="1"
+                                />
+                                <field
+                                    name="product_uom"
+                                    groups="uom.group_uom"
+                                    class="oe_inline"
+                                    readonly="1"
+                                />
+                            </div>
+                            <label for="remaining_uom_qty" />
+                            <div>
+                                <field
+                                    name="remaining_uom_qty"
+                                    class="oe_inline"
+                                    readonly="1"
+                                />
+                                <field
+                                    name="product_uom"
+                                    groups="uom.group_uom"
+                                    class="oe_inline"
+                                    readonly="1"
+                                />
+                            </div>
+                            <label for="remaining_qty" />
+                            <div>
+                                <field
+                                    name="remaining_qty"
+                                    class="oe_inline"
+                                    readonly="1"
+                                />
+                            </div>
+                        </group>
+                    </group>
+                    <notebook>
+                        <page name="sale_lines" string="Sale Order Lines">
+                            <field
+                                name="sale_lines"
+                                mode="tree"
+                                readonly="1"
+                                domain="[('product_id', '=', product_id)]"
+                            />
+                        </page>
+                    </notebook>
+                </sheet>
+                <div class="oe_chatter">
+                    <field name="message_follower_ids" widget="mail_followers" />
+                    <field name="activity_ids" widget="mail_activity" />
+                    <field name="message_ids" widget="mail_thread" />
+                </div>
+            </form>
+        </field>
+    </record>
+
+    <record id="sale_blanket_order_line_search" model="ir.ui.view">
+        <field name="name">sale.blanket.order.line.search</field>
+        <field name="model">sale.blanket.order.line</field>
+        <field name="arch" type="xml">
+            <search string="Search Sale Blanket Order Line">
+                <field name="product_id" />
+                <field name="date_schedule" />
+            </search>
+        </field>
+    </record>
+
+    <!-- ACTIONS (SERVER) -->
+
+    <record
+        id="act_open_sale_blanket_order_lines_view_tree"
+        model="ir.actions.act_window"
+    >
+        <field name="name">Blanket Order Lines</field>
+        <field name="type">ir.actions.act_window</field>
+        <field name="res_model">sale.blanket.order.line</field>
+        <field name="view_mode">tree,form</field>
+        <field name="search_view_id" ref="sale_blanket_order_line_search" />
+    </record>
+
+    <!-- MENU'S -->
+
+    <menuitem
+        action="act_open_sale_blanket_order_lines_view_tree"
+        id="menu_sale_blanket_order_line"
+        groups="sales_team.group_sale_salesman"
+        sequence="21"
+        parent="sale.sale_order_menu"
+    />
+
+</odoo>
