+++ sale_blanket_order/views/sale_order_views.xml
@@ -20,13 +20,11 @@
         <field name="name">sale.order.from.blanket.form - disable adding lines</field>
         <field name="model">sale.order</field>
         <field name="inherit_id" ref="sale.view_order_form" />
-        <field
-            name="groups_id"
-            eval="[(6,0,[ref('sale_blanket_order.blanket_orders_disable_adding_lines')])]"
-        />
         <field name="arch" type="xml">
             <xpath expr="//field[@name='order_line']//tree" position="attributes">
-                <attribute name="create">blanket_order_id==False</attribute>
+                <t groups="sale_blanket_order.blanket_orders_disable_adding_lines">
+                    <attribute name="create">blanket_order_id==False</attribute>
+                </t>
             </xpath>
         </field>
     </record>
