+++ setup/sale_blanket_order/odoo/addons/sale_blanket_order
@@ -1,0 +1,1 @@
+../../../../sale_blanket_order
